+++ ui/accordion/src/AccordionContent.tsx
@@ -21,9 +21,15 @@ const AnimatedContent = styled(AccordionPrimitive.Content, {
   color: theme.colors.primary,
   '&[data-state="open"]': {
     animation: `${slideDown} ${theme.transitions.normal} ${easeInOutExpo}`,
+    "@reducedMotion": {
+      animation: "none",
+    },
   },
   '&[data-state="closed"]': {
     animation: `${slideUp} ${theme.transitions.normal} ${easeInOutExpo}`,
+    "@reducedMotion": {
+      animation: "none",
+    },
   },
 });