+++ ui/carousel/src/CarouselContent.tsx
@@ -22,6 +22,9 @@ const Container = styled("div", {
 const Slider = styled("div", {
   display: "flex",
   transition: `transform 0.5s ${theme.transitions.inOut}`,
+  "@reducedMotion": {
+    transition: "none",
+  },
 });
 
 export type CarouselContentProps = {